requirements.txt
Django==1.11.11
djangorestframework==3.7.7
djangorestframework-jwt==1.11.0
psycopg2==2.7.4
PyJWT==1.6.1
pytz==2018.3
